--- src/main/java/com/henry/tryout/easy_coding/jvm/bytecode_and_classloader_01/Another.java
@@ -1,7 +1,0 @@
-package com.henry.tryout.easy_coding.jvm.bytecode_and_classloader_01;
-
-public class Another {
-    public void speak() {
-        System.out.println("easy coding.");
-    }
-}
