--- src/main/java/com/henry/tryout/easy_coding/jvm/bytecode_and_classloader_01/CustomClassLoader_05.java
@@ -1,37 +1,0 @@
-package com.henry.tryout.easy_coding.jvm.bytecode_and_classloader_01;
-
-import java.io.FileNotFoundException;
-
-public class CustomClassLoader_05 extends ClassLoader {
-
-    @Override
-    protected Class<?> findClass(String name) throws ClassNotFoundException {
-        try {
-            byte[] result = getClassFromCustomPath(name);
-            if (result == null) {
-                throw new FileNotFoundException();
-            } else {
-                return defineClass(name, result, 0, result.length);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        throw new ClassNotFoundException(name);
-    }
-
-    private byte[] getClassFromCustomPath(String name) {
-        // 从自定义的路径中加载 类
-        return null;
-    }
-
-    public static void main(String[] args) {
-        CustomClassLoader_05 customClassLoader = new CustomClassLoader_05();
-        try {
-            Class<?> clazz = Class.forName("One", true, customClassLoader);
-            Object obj = clazz.newInstance();
-            System.out.println(obj.getClass().getClassLoader());
-        } catch (ClassNotFoundException | InstantiationException | IllegalAccessException e) {
-            e.printStackTrace();
-        }
-    }
-}
