--- src/main/java/com/henry/tryout/easy_coding/jvm/bytecode_and_classloader_01/One.java
@@ -1,13 +1,0 @@
-package com.henry.tryout.easy_coding.jvm.bytecode_and_classloader_01;
-
-public class One {
-    private String inner = "time files.";
-
-    public void call() {
-        System.out.println("Hello world.");
-    }
-
-    public String getInner() {
-        return inner;
-    }
-}
