+++ src/main/java/com/henry/tryout/easy_coding/jvm_04/bytecode_and_classloader_01/bytecode_command_00
@@ -1,0 +1,50 @@
+程序代码 -> 字节码(中间层) -> CPU指令
+
+一个字节就能够表示 Java中所有存在的指令。
+Java指令总共有200个，而一个字节可以表示 256种可能
+
+cafe babe : 标识文件是一个Java类文件；
+
+字节码的助记符号 👇
+特征：首字母表示 变量的类型。 A表示引用类型变量， I表示int类型变量
+字节码指令分类：
+    1 加载/存储指令；
+                      加载到
+    虚拟机栈的 局部变量表 <-> 操作栈
+                      存储到
+      分类：
+        1 把 局部变量 加载到 操作栈中； ILOAD
+        2 把 操作栈顶 存储到 局部变量表； - ISTORE
+        3 把 常量 加载到 操作栈顶； ICONST等
+
+    2 运算指令；
+        用来 对操作栈上的两个值进行运算，并把结果写回到 操作栈顶。
+        ISTORE
+
+    3 类型转换指令；
+        作用：显示转换 两个不同的数值类型
+
+    4 对象的创建与访问指令；
+        分类：
+            1 创建对象的指令； - NEW
+            2 访问属性的指令； - GETFIELD
+            3 检查实例类型的指令； - INSTANCEOF
+
+    5 操作栈管理指令；
+        作用：直接控制 操作栈；
+        分类：
+            1 出栈操作；POP
+            2 复制 栈顶元素 并 压入栈；DUP
+
+    6 方法调用与返回指令；
+        分类：
+            1 调用对象的实例方法 - INVOKEVIRTUAL;
+            2 调用实例的初始化方法、私有方法、父类方法 - INVOKESPECIAL
+            3 调用 类的静态方法 - INVOKESTATIC
+            4 返回Void类型 - RETURN
+
+    7 同步指令；
+        作用：支持语法层面的同步操作
+        用法：
+            1 标记方法为同步方法； - ACC_SYNCHRONIZED
+            2 支持 synchronized语义 - MONITORENTER + MONITEREXIT
