+++ src/main/java/com/henry/tryout/easy_coding/jvm_04/memory_layout_02/memory_district
